# Optimize threshold of BigInteger.Parse