--- a/src/libraries/System.Runtime.Numerics/src/System/Number.BigInteger.cs
+++ b/src/libraries/System.Runtime.Numerics/src/System/Number.BigInteger.cs
@@ -668,7 +668,7 @@ namespace System
 #else
         internal const
 #endif
-        int s_naiveThreshold = 20000;
+        int s_naiveThreshold = 3200;
         private static ParsingStatus NumberToBigInteger(ref NumberBuffer number, out BigInteger result)
         {
             int currentBufferSize = 0;